Move CFG printing from optimizer to `ASTContext`.
