+++ hilti/toolchain/src/ast/ast-context.cc
@@ -18,6 +18,7 @@
 #include <hilti/ast/visitor.h>
 #include <hilti/base/timing.h>
 #include <hilti/compiler/detail/ast-dumper.h>
+#include <hilti/compiler/detail/cfg.h>
 #include <hilti/compiler/detail/optimizer.h>
 #include <hilti/compiler/detail/resolver.h>
 #include <hilti/compiler/detail/scope-builder.h>
@@ -40,6 +41,9 @@ inline const DebugStream AstResolved("ast-resolved");
 inline const DebugStream AstTransformed("ast-transformed");
 inline const DebugStream Compiler("compiler");
 inline const DebugStream Resolver("resolver");
+inline const DebugStream CfgInitial("cfg-initial");
+inline const DebugStream CfgFinal("cfg-final");
+
 } // namespace hilti::logging::debug
 
 namespace hilti::ast::detail {
@@ -780,6 +784,9 @@ Result<Nothing> ASTContext::_transform(Builder* builder, const Plugin& plugin) {
 }
 
 Result<Nothing> ASTContext::_optimize(Builder* builder) {
+    if ( logger().isEnabled(logging::debug::CfgInitial) )
+        hilti::detail::cfg::dump(logging::debug::CfgInitial, _root);
+
     HILTI_DEBUG(logging::debug::Compiler, "performing global transformations");
 
     bool first = true;
@@ -796,6 +803,9 @@ Result<Nothing> ASTContext::_optimize(Builder* builder) {
         if ( auto rc = _resolve(builder, plugin::registry().hiltiPlugin()); ! rc )
             return rc;
     }
+
+    if ( logger().isEnabled(logging::debug::CfgFinal) )
+        hilti::detail::cfg::dump(logging::debug::CfgFinal, _root);
 
     // Make sure we didn't leave anything odd during optimization.
     _checkAST(true);
